+++ opensearchapi/api_component_template.go
@@ -22,7 +22,7 @@ func (c componentTemplateClient) Create(ctx context.Context, req ComponentTempla
 		data ComponentTemplateCreateResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, c.apiClient, req, &data); err != nil {
 		return &data, err
 	}
 
@@ -35,7 +35,7 @@ func (c componentTemplateClient) Delete(ctx context.Context, req ComponentTempla
 		data ComponentTemplateDeleteResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, c.apiClient, req, &data); err != nil {
 		return &data, err
 	}
 
@@ -52,7 +52,7 @@ func (c componentTemplateClient) Get(ctx context.Context, req *ComponentTemplate
 		data ComponentTemplateGetResp
 		err  error
 	)
-	if data.response, err = c.apiClient.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, c.apiClient, req, &data); err != nil {
 		return &data, err
 	}
 
@@ -61,5 +61,5 @@ func (c componentTemplateClient) Get(ctx context.Context, req *ComponentTemplate
 
 // Exists executes a exists componentTemplate request with the required ComponentTemplatExistsReq
 func (c componentTemplateClient) Exists(ctx context.Context, req ComponentTemplateExistsReq) (*opensearch.Response, error) {
-	return c.apiClient.do(ctx, req, nil)
+	return doRequest(ctx, c.apiClient, req)
 }