+++ opensearchapi/api_mget.go
@@ -22,7 +22,7 @@ func (c Client) MGet(ctx context.Context, req MGetReq) (*MGetResp, error) {
 		data MGetResp
 		err  error
 	)
-	if data.response, err = c.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, &c, req, &data); err != nil {
 		return &data, err
 	}