+++ opensearchapi/api_render_search_template.go
@@ -22,7 +22,7 @@ func (c Client) RenderSearchTemplate(ctx context.Context, req RenderSearchTempla
 		data RenderSearchTemplateResp
 		err  error
 	)
-	if data.response, err = c.do(ctx, req, &data); err != nil {
+	if data.response, err = do(ctx, &c, req, &data); err != nil {
 		return &data, err
 	}